feat: implement QuestionBuilder component for dynamic question creation
